+++ lib/src/unit/value_range.dart
@@ -7,7 +7,7 @@ part 'value_range.freezed.dart';
 @freezed
 class ValueRange<V, U extends Unit> with _$ValueRange<V, U> {
   const factory ValueRange({
-    ValueUnit<V, U>? start,
-    ValueUnit<V, U>? end,
+    required ValueUnit<V, U> start,
+    required ValueUnit<V, U> end,
   }) = _ValueRange<V, U>;
 }