+++ lib/src/unit/value_unit.dart
@@ -9,29 +9,11 @@ class ValueUnit<V, U extends Unit> with _$ValueUnit<V, U> {
   const ValueUnit._();
 
   const factory ValueUnit({
-    V? value,
-    U? unit,
+    required V value,
+    required U unit,
   }) = _ValueUnit<V, U>;
 
-  String toCompactString() {
-    if (value != null) {
-      String displayString = value!.toString();
-      if (unit != null) {
-        displayString += unit!.symbol;
-      }
-      return displayString;
-    }
-    return '';
-  }
+  String toCompactString() => '$value${unit.symbol}';
 
-  String toDisplayString() {
-    if (value != null) {
-      String displayString = value!.toString();
-      if (unit != null) {
-        displayString += ' ${unit!.symbol}';
-      }
-      return displayString;
-    }
-    return '';
-  }
+  String toDisplayString() => '$value ${unit.symbol}';
 }