+++ lib/src/unit/velocity.g.dart
@@ -1,0 +1,16 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'velocity.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+VelocityUnit _$VelocityUnitFromJson(Map<String, dynamic> json) => VelocityUnit(
+      symbol: json['symbol'] as String,
+    );
+
+Map<String, dynamic> _$VelocityUnitToJson(VelocityUnit instance) =>
+    <String, dynamic>{
+      'symbol': instance.symbol,
+    };